+++ examples/python.py
@@ -1,4 +1,4 @@
-from pyinfra.modules import python
+from pyinfra.operations import python
 
 # Tip: Can run try it out using: 'pyinfra @docker/python python.py'