+++ pyinfra/modules.py
@@ -1,0 +1,7 @@
+from pyinfra import logger
+from pyinfra.operations import *  # noqa: F401,F403
+from pyinfra.operations import __all__
+
+__all__ = __all__
+
+logger.warning('Use of `pyinfra.operations` is deprecated, please use `pyinfra.operations`.')